[QLN-67] feat: add form to copy dialog
