+++ src/app/[locale]/(main)/quizbank/[id]/components/author-quizbank.tsx
@@ -9,7 +9,7 @@ import { useTranslations } from "next-intl"
 import { Avatar, AvatarFallback, AvatarImage } from "@/components/ui/avatar"
 import { getShortName } from "@/lib/string-helper"
 import CopyButton from "@/components/copy-button"
-import CopyQuizBankDialog from "./copy-quizbank-dialog"
+import CopyQuizBankDialog from "./copy-quizbank-dialog/copy-quizbank-dialog"
 
 type Props = {
   authorData: any
